pyproject.toml
[project]
name = "mcp2plugin"
version = "0.1.0"
description = "Convert MCP servers to Claude Code plugins"
readme = "README.md"
authors = [
    { name = "Kedia", email = "howardsue319@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "google-genai>=1.58.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[project.scripts]
mcp2plugin = "mcp2plugin.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
